fix: remove dependency to util-linux column and fix column splitting
Use portable options of column (-s and -t).
We now indent on the tab character instead of whitespaces, so the column alignement does not break if paths or filenames contain whitespaces.
We provide a column separator placeholder that is replaced at runtime by starlist.

Also added new placeholders for the colors, to be able to reflect a change in the colors without having to export the list command again.

Reflected the change in README and in tests.